Errores con Funciones y tipos
e

+++ Proyecto1_Subastas_Postgresql/src/Controller/PujaController.java
@@ -32,7 +32,7 @@ public class PujaController {
     }
     
 
-    public boolean insertarPuja(double compradorId, double subastaId, double precio, ArrayList<Puja> listaPuja) throws ParseException, SQLException{
+    public boolean insertarPuja(int compradorId, int subastaId, double precio, ArrayList<Puja> listaPuja) throws ParseException, SQLException{
  
        Puja Puja = new Puja(compradorId,subastaId,precio);
        System.out.println("antes del if en puja controller");
